A new hope but I have another idea. Be nice if I could pass a env flag to main
